Intermediate commit
Adding the products of the gating vector with the activations yields their respective effect on the final activation. However that effect is typically rather small. Thus I will convert the code back to the simple vector representations. That mean effect will be larger but it might be less clear how the vector affect the activations vector. Will be subject of discussion.